refactor: Numbers 생성 유효성 검사 테스트 코드 NumbersTest 로 이관
